@@ -1,11 +1,11 @@
 name: fotoapparat
 description: Bir kamera plugindir.
-version: 0.0.2
+version: 0.0.3
 author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
